+++ git-index/src/decode/mod.rs
@@ -204,6 +204,7 @@ impl State {
             tree,
             link,
             resolve_undo,
+            untracked,
             is_sparse: is_sparse_from_ext, // a marker is needed in case there are no directories
         } = ext;
         is_sparse |= is_sparse_from_ext;
@@ -219,6 +220,7 @@ impl State {
                 tree,
                 link,
                 resolve_undo,
+                untracked,
             },
             checksum,
         ))
