+++ git-index/tests/index/file/mod.rs
@@ -81,6 +81,13 @@ mod init {
     }
 
     #[test]
+    #[ignore]
+    fn read_untr_extension() {
+        let file = loose_file("UNTR");
+        assert_eq!(file.version(), Version::V2);
+    }
+
+    #[test]
     fn read_v4_with_delta_paths_and_ieot_ext() {
         let file = file("v4_more_files_IEOT");
         assert_eq!(file.version(), Version::V4);
